[refactor] Photographer 서비스 구조 개선 / 주변 작가 조회 필터링
