+++ src/main/java/com/hm/picplz/domain/photographer/controller/DefaultCameraController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.hm.picplz.domain.photographer.dto.DefaultCameraDto;
-import com.hm.picplz.domain.photographer.service.DefaultCameraService;
+import com.hm.picplz.domain.photographer.service.PhotographerService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -21,11 +21,11 @@ import lombok.extern.slf4j.Slf4j;
 @Tag(name = "DefaultCamera")
 public class DefaultCameraController {
 
-	private final DefaultCameraService defaultCameraService;
+	private final PhotographerService photographerService;
 
 	@Operation(summary = "기획서에서 정한 기본 촬영 기기 목록 반환")
 	@GetMapping
 	public List<DefaultCameraDto.CameraInfo> getAllCameras() {
-		return defaultCameraService.getCameras();
+		return photographerService.getCameras();
 	}
 }
